# api-gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Smart City Management System resources",
        "title": "Smart City Management System API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/citizens/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Register a new citizen with the provided details.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "citizens"
                ],
                "summary": "Create a new citizen",
                "parameters": [
                    {
                        "description": "Citizen Data",
                        "name": "citizen",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/citizen.CitizenCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/citizen.Citizen"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/citizens/{citizenId}/documents": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create documents for a specific citizen.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "citizens"
                ],
                "summary": "Create documents for a citizen",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Citizen ID",
                        "name": "citizenId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Document Data",
                        "name": "document",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/citizen.ById"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/citizen.DocumentList"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/citizens/{citizenId}/getbyid": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get the profile details of a citizen by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "citizens"
                ],
                "summary": "Get citizen profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Citizen ID",
                        "name": "citizenId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/citizen.Citizen"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/citizens/{citizenId}/requests": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all service requests made by a specific citizen.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "citizens"
                ],
                "summary": "Get citizen requests",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Citizen ID",
                        "name": "citizenId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/citizen.ServiceReqList"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new service request for a citizen.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "citizens"
                ],
                "summary": "Create a new service request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Citizen ID",
                        "name": "citizenId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Service Request Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/citizen.ServiceReqCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/citizen.ServiceReq"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/citizens/{citizenId}/update": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update the profile details of a citizen by ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "citizens"
                ],
                "summary": "Update citizen profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Citizen ID",
                        "name": "citizenId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated Citizen Data",
                        "name": "citizen",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/citizen.CitizenCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/citizen.Citizen"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "citizen.ById": {
            "type": "object",
            "properties": {
                "Id": {
                    "type": "string"
                }
            }
        },
        "citizen.Citizen": {
            "type": "object",
            "properties": {
                "Address": {
                    "type": "string"
                },
                "CitizenId": {
                    "type": "string"
                },
                "CreateAt": {
                    "type": "string"
                },
                "DateOfBirth": {
                    "type": "string"
                },
                "DeleteAt": {
                    "type": "string"
                },
                "Email": {
                    "type": "string"
                },
                "FirstName": {
                    "type": "string"
                },
                "LastName": {
                    "type": "string"
                },
                "PhoneNumber": {
                    "type": "string"
                },
                "UpdateAt": {
                    "type": "string"
                },
                "UserId": {
                    "type": "string"
                }
            }
        },
        "citizen.CitizenCreate": {
            "type": "object",
            "properties": {
                "Address": {
                    "type": "string"
                },
                "CitizenId": {
                    "type": "string"
                },
                "DateOfBirth": {
                    "type": "string"
                },
                "Email": {
                    "type": "string"
                },
                "FirstName": {
                    "type": "string"
                },
                "LastName": {
                    "type": "string"
                },
                "PhoneNumber": {
                    "type": "string"
                },
                "UserId": {
                    "type": "string"
                }
            }
        },
        "citizen.CitizenModel": {
            "type": "object",
            "properties": {
                "Address": {
                    "type": "string"
                },
                "DateOfBirth": {
                    "type": "string"
                },
                "Email": {
                    "type": "string"
                },
                "FirstName": {
                    "type": "string"
                },
                "LastName": {
                    "type": "string"
                },
                "PhoneNumber": {
                    "type": "string"
                },
                "UserId": {
                    "type": "string"
                }
            }
        },
        "citizen.Document": {
            "type": "object",
            "properties": {
                "CitizenId": {
                    "type": "string"
                },
                "CreateAt": {
                    "type": "string"
                },
                "DeleteAt": {
                    "type": "string"
                },
                "DocumentId": {
                    "type": "string"
                },
                "DocumentType": {
                    "type": "string"
                },
                "FilePath": {
                    "type": "string"
                },
                "UpdateAt": {
                    "type": "string"
                }
            }
        },
        "citizen.DocumentList": {
            "type": "object",
            "properties": {
                "Model": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/citizen.Document"
                    }
                }
            }
        },
        "citizen.Service": {
            "type": "object",
            "properties": {
                "CitizenId": {
                    "$ref": "#/definitions/citizen.CitizenModel"
                },
                "CreateAt": {
                    "type": "string"
                },
                "Description": {
                    "type": "string"
                },
                "RequestType": {
                    "type": "string"
                },
                "Status": {
                    "type": "string"
                },
                "UpdateAt": {
                    "type": "string"
                }
            }
        },
        "citizen.ServiceReq": {
            "type": "object",
            "properties": {
                "CitizenId": {
                    "type": "string"
                },
                "CreateAt": {
                    "type": "string"
                },
                "DeleteAt": {
                    "type": "string"
                },
                "Description": {
                    "type": "string"
                },
                "RequestId": {
                    "type": "string"
                },
                "RequestType": {
                    "type": "string"
                },
                "Status": {
                    "type": "string"
                },
                "UpdateAt": {
                    "type": "string"
                }
            }
        },
        "citizen.ServiceReqCreate": {
            "type": "object",
            "properties": {
                "CitizenId": {
                    "type": "string"
                },
                "Description": {
                    "type": "string"
                },
                "RequestId": {
                    "type": "string"
                },
                "RequestType": {
                    "type": "string"
                },
                "Status": {
                    "type": "string"
                }
            }
        },
        "citizen.ServiceReqList": {
            "type": "object",
            "properties": {
                "Count": {
                    "type": "integer"
                },
                "Model": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/citizen.Service"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}